BACKPORT: FROMGIT: sched/deadline: Account idle time for deferred server
In the rt throttling mechanism, the real-time schedulers runtime is
accounted while the rt schedulers run. This implies that idle
time is also considered time available for non-RT tasks.

To avoid regressions, consider the idle time as time available for the
fair server if the server is of defer mode (iow compatible with rt
throttling).

For the non-defer mode, this is not a problem, as the server aims
explicitly to provide bandwidth for fair tasks, preempting the rt
schedulers without deffering.

Signed-off-by: Daniel Bristot de Oliveira <bristot@kernel.org>

(cherry picked from commit 46a04dc08aec247d43247d411da82c4c78d048c5
 https://git.kernel.org/pub/scm/linux/kernel/git/bristot/linux.git dl_server_v6)

Conflicts:
	kernel/sched/deadline.c
	kernel/sched/fair.c
	kernel/sched/idle.c

BUG=b:307273029
TEST=booted

Change-Id: I3c96e35f307c5e6d21aca1ed9d3da21e68ce82a8
Signed-off-by: Suleiman Souhlal <suleiman@google.com>